[FIX] LockSheetPlugin: move the plugin to the correct registry
The LockSheet plugin does not have a state to synchronize. Its sole
purpose is to prevent the user from modifying a locked sheet at the UI
level.
As such, it is a UiFeature plugin.

Task: 6428105
