+++ src/types/getters.ts
@@ -15,6 +15,7 @@ import { ColorThemeUIPlugin } from "../plugins/ui_feature/color_theme";
 import { DynamicTranslate } from "../plugins/ui_feature/dynamic_translate";
 import { GeoFeaturePlugin } from "../plugins/ui_feature/geo_features";
 import { HistoryPlugin } from "../plugins/ui_feature/local_history";
+import { LockSheetPlugin } from "../plugins/ui_feature/lock_sheet";
 import { PivotPresencePlugin } from "../plugins/ui_feature/pivot_presence_plugin";
 import { SortPlugin } from "../plugins/ui_feature/sort";
 import { SubtotalEvaluationPlugin } from "../plugins/ui_feature/subtotal_evaluation";
@@ -27,7 +28,6 @@ import { FigureUIPlugin } from "../plugins/ui_stateful/figure";
 import { FilterEvaluationPlugin } from "../plugins/ui_stateful/filter_evaluation";
 import { HeaderPositionsUIPlugin } from "../plugins/ui_stateful/header_positions";
 import { HeaderVisibilityUIPlugin } from "../plugins/ui_stateful/header_visibility_ui";
-import { LockSheetPlugin } from "../plugins/ui_stateful/lock_sheet";
 import { GridSelectionPlugin } from "../plugins/ui_stateful/selection";
 import { TableComputedStylePlugin } from "../plugins/ui_stateful/table_computed_style";
 import { CoreGetters, PluginGetters } from "./core_getters";
